Refactor import statements for improved module organization and consistency
