+++ processmine/processmine/tests/integration_test.py
@@ -77,7 +77,7 @@ class ProcessMineIntegrationTests(unittest.TestCase):
         
         # Check if package is available
         try:
-            import processmine.processmine as processmine
+            import processmine
             cls.is_available = True
         except ImportError:
             cls.is_available = False
@@ -94,7 +94,8 @@ class ProcessMineIntegrationTests(unittest.TestCase):
             self.skipTest("ProcessMine package not available")
         
         # Import necessary modules
-        from processmine.processmine import create_model, run_analysis
+        from processmine.models.factory import create_model
+        from processmine.core.runner import run_analysis
         from processmine.data.loader import load_and_preprocess_data
         from processmine.data.graphs import build_graph_data
         from processmine.core.training import train_model, evaluate_model
@@ -152,7 +153,7 @@ class ProcessMineIntegrationTests(unittest.TestCase):
             test_loader = DataLoader([graphs[i] for i in test_idx], batch_size=4)
             
             # Step 4: Create model
-            from processmine.processmine import create_model
+            from processmine.models.factory import create_model
             
             model = create_model(
                 model_type="enhanced_gnn",
@@ -269,7 +270,7 @@ class ProcessMineIntegrationTests(unittest.TestCase):
     
     def test_cli_integration(self):
         """Test the command-line interface."""
-        from processmine.processmine.cli import main, parse_arguments
+        from processmine.cli import main, parse_arguments
         
         # Test with analyze mode
         test_args = [
@@ -311,7 +312,7 @@ class ProcessMineIntegrationTests(unittest.TestCase):
     
     def test_model_types_integration(self):
         """Test creating and using different model types."""
-        from processmine.processmine import create_model
+        from processmine.models.factory import create_model
         
         # Load and preprocess data
         from processmine.data.loader import load_and_preprocess_data
